Add playback controls for SamplePad
The old global playback controls have been replaced with playback
controls that are sample-specific. The controls use Material-UI icons,
at the moment only looping and "play" is supported. The "bars" control
is a placeholder that's intended for specifying the number of bars that
a sample will be repeated for.

that are sample-specific for playback and looping. At the moment