Add synth triggers to Sequencer
This commit abstracts the grid from the Sequencer into a SequencerGrid
component. There's also a new Arrangement class which at the moment
maintains a matrix that represents the internal state of the sequencer
grid. Whenever a sequencer cell is selected:
  1. Arrangement matrix gets updated.
  2. The React component's internal state gets updated with a
     clone of the updated arrangement matrix.
  3. This triggers a rerender of the component.
There's only one instance of the Arrangement and it's being queried as
part of the sequencer step callback.

The work for the component is still in progress. It needs controls for
Start & Play as well, the triggering of different samples instead of
synthesized notes and indication of the current beat.