Fix slider scaling issues
The FxParamSlider is now being scaled in accordance to the paramer's
value range. This also simplifies the code that's used to set fx
parameter values and gets rid of the readonly params. We can not
adjust the params of a connected fx node, therefore it gets recreated
every time the slider commits a value.