Move to pre-16th scheduling
The previous way of scheduling involved scheduling for a specific time
ahead, e.g. start of measure 6. While brilliant on paper, in practice
it didn't work as the target time would end up being highly inaccurate.
The inaccuracies would become more and more clear if a loop was present
for a long time as the deviations from rhythm would keep on increasing.

The fix for that was moving to a system that triggers samples right
before they are meant to be played.

This also introduces a modular "Start" dialog for starting playback.