Wire up effects
The effects can now be triggered using FxPad.

The SampleTable has been removed as it is redundant. It was used to
sketch out the initial implementation.

New FxTrigger class that represents the data required to apply an
effect on a sample. I left the other Fx classes in just in case they
turn out to be useful for testing out effects using the FxPanel.