Fix image preload issue
The images weren't getting preloaded on Chrome because it would think
that they're not required if not used immediately. Making them
invisible and dumping them onto DOM fixed the issue.

The http-server also had a caching setting of -c-1 which is unsuitable
as we want to cache the images at least for some time to minimise data
usage.