Fix UI scale bug in XYPadStyled
The problem was that whenever we did a transform when the window size
changed it would throw off the hit box of the XYPad control region.
It turns out that x/y adjustments have to be applied in the original
coordinate space as opposed to the scaled one. The fix is a dirty one.
We grab the UI scale from the transform property of the main container
and apply an inverse transform to XYPad calculations.